remove deasync, move register to seperate package
