+++ packages/core/core/package.json
@@ -35,7 +35,6 @@
     "json5": "^1.0.1",
     "micromatch": "^4.0.2",
     "nullthrows": "^1.1.1",
-    "pirates": "^4.0.0",
     "semver": "^5.4.1"
   },
   "devDependencies": {
