+++ packages/core/register/package.json
@@ -22,6 +22,8 @@
     "@parcel/config-default": "^2.0.0-alpha.3.1",
     "@parcel/core": "^2.0.0-alpha.3.2",
     "@parcel/fs": "^2.0.0-alpha.3.1",
-    "@parcel/package-manager": "^2.0.0-alpha.3.1"
+    "@parcel/package-manager": "^2.0.0-alpha.3.1",
+    "pirates": "^4.0.0",
+    "deasync": "^0.1.14"
   }
 }
