+++ packages/core/register/src/syncPromise.js
