+++ packages/core/utils/package.json
@@ -23,7 +23,6 @@
     "@parcel/markdown-ansi": "^2.0.0-alpha.3.1",
     "ansi-html": "^0.0.7",
     "clone": "^2.1.1",
-    "deasync": "^0.1.14",
     "fast-glob": "^3.0.4",
     "is-glob": "^4.0.0",
     "is-url": "^1.2.2",
